+++ app/Drivers.php
@@ -6,5 +6,8 @@ use Illuminate\Database\Eloquent\Model;
 
 class Drivers extends Model
 {
+	public function route(){
+		return $this->hasMany(Routes::class, 'DriverID');
+	}
     //
 }